Add tests from Xamarin Studio.
