+++ .gitignore
@@ -38,3 +38,7 @@ Thumbs.db
 
 # dotCover
 *.dotCover
+
+# Test runner
+testrunner/
+TestResult.xml
