+++ CSharpTraining.sln
@@ -9,9 +9,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		.travis.yml = .travis.yml
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CSharpTraining.Tests", "CSharpTraining.Tests\CSharpTraining.Tests.csproj", "{626DE17F-6626-4133-A1FD-CA52D01452A0}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MonoBrickFirmware", "monoev3\MonoBrickFirmware\MonoBrickFirmware.csproj", "{67261E03-D263-4C42-A5AD-2A4820231B28}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NUnitSample.Tests", "NUnitSample.Tests\NUnitSample.Tests.csproj", "{F12663F4-5243-492E-95AC-48DB7D137689}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -29,14 +29,6 @@ Global
 		{4B4D93F6-AB41-4404-A615-E484752886BE}.Release|Any CPU.Build.0 = Release|Any CPU
 		{4B4D93F6-AB41-4404-A615-E484752886BE}.Release|x86.ActiveCfg = Release|x86
 		{4B4D93F6-AB41-4404-A615-E484752886BE}.Release|x86.Build.0 = Release|x86
-		{626DE17F-6626-4133-A1FD-CA52D01452A0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{626DE17F-6626-4133-A1FD-CA52D01452A0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{626DE17F-6626-4133-A1FD-CA52D01452A0}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{626DE17F-6626-4133-A1FD-CA52D01452A0}.Debug|x86.Build.0 = Debug|Any CPU
-		{626DE17F-6626-4133-A1FD-CA52D01452A0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{626DE17F-6626-4133-A1FD-CA52D01452A0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{626DE17F-6626-4133-A1FD-CA52D01452A0}.Release|x86.ActiveCfg = Release|Any CPU
-		{626DE17F-6626-4133-A1FD-CA52D01452A0}.Release|x86.Build.0 = Release|Any CPU
 		{67261E03-D263-4C42-A5AD-2A4820231B28}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{67261E03-D263-4C42-A5AD-2A4820231B28}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{67261E03-D263-4C42-A5AD-2A4820231B28}.Debug|x86.ActiveCfg = Debug|Any CPU
@@ -45,6 +37,14 @@ Global
 		{67261E03-D263-4C42-A5AD-2A4820231B28}.Release|Any CPU.Build.0 = Release|Any CPU
 		{67261E03-D263-4C42-A5AD-2A4820231B28}.Release|x86.ActiveCfg = Release|Any CPU
 		{67261E03-D263-4C42-A5AD-2A4820231B28}.Release|x86.Build.0 = Release|Any CPU
+		{F12663F4-5243-492E-95AC-48DB7D137689}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F12663F4-5243-492E-95AC-48DB7D137689}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F12663F4-5243-492E-95AC-48DB7D137689}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{F12663F4-5243-492E-95AC-48DB7D137689}.Debug|x86.Build.0 = Debug|Any CPU
+		{F12663F4-5243-492E-95AC-48DB7D137689}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F12663F4-5243-492E-95AC-48DB7D137689}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F12663F4-5243-492E-95AC-48DB7D137689}.Release|x86.ActiveCfg = Release|Any CPU
+		{F12663F4-5243-492E-95AC-48DB7D137689}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
