+++ NUnitSample.Tests/packages.config
@@ -1,5 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="NUnit" version="2.6.4" targetFramework="net45" />
-  <package id="NUnit.Runners" version="2.6.4" targetFramework="net45" />
 </packages>
